Created tests and confirmed functionality

+++ inc/temperature.h
@@ -1,1 +1,1 @@
-long bmp180_get_temperature(struct BMP180_CALIBRATION *, long);
+long bmp180_get_temperature(struct BMP180_CALIBRATION *, int16_t);
